refactor(backend): rename pull-requests.dto to pr-list.dto and export GraphQLPRNode type
